@@ -9,7 +9,9 @@
 		"homepage": "http://dreamysource.io/project/funktions"
 	}],
 	"require": {
-		"php": ">=5.3.0"
+		"php": ">=5.4.0",
+        "ext-curl": "*",
+        "ext-fileinfo": "*"
 	},
 	"autoload": {
 		"files": [
